+++ lib/auto_form_validate.dart
@@ -4,3 +4,4 @@ export 'package:flutter/src/services/text_input.dart';
 export 'auto_text_form_field.dart';
 export 'package:mask_text_input_formatter/mask_text_input_formatter.dart';
 export 'form_controller.dart';
+export 'auto_form_field_wrapper.dart';